Completed ts convervsion of db and json db
