--- app/db/connect_pool.js
@@ -1,9 +1,0 @@
-import pg from 'pg';
-
-export const createPool = () => (new pg.Pool({
-        host: process.env.DB_HOST,
-        database: process.env.DB,
-        user: process.env.DB_USER,
-        password: process.env.DB_PASSWORD,
-        port: parseInt(process.env.DB_PORT)
-}));
