+++ app/db/index.ts
@@ -1,0 +1,7 @@
+import createPool from './createPool'; 
+import runQuery from './runQuery'; 
+
+export {
+    createPool,
+    runQuery
+}
