+++ package.json
@@ -14,6 +14,7 @@
   "devDependencies": {
     "@types/node": "^22.10.3",
     "@types/node-fetch": "^2.6.12",
+    "@types/pg": "^8.11.10",
     "@types/webpack-node-externals": "^3.0.4",
     "babel-loader": "^9.2.1",
     "ts-loader": "^9.5.1",
